Lab 6 work

+++ lab5/main.cpp
@@ -2,7 +2,7 @@
 
 double ctok(double c)
 {
-    int k = c + 273.15;
+    double k = c + 273.15;
     return k;
 }
